requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.1
langchain-core==0.3.6
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.0
langsmith==0.1.129
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
narwhals==1.8.4
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.5
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pygame==2.6.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uv==0.4.17
watchdog==4.0.2
yarl==1.13.1
loguru==0.7.2
